move events out of the rooms.
