+++ internal/api/api.go
@@ -61,6 +61,11 @@ func (manager *ApiManagerCtx) Mount(r chi.Router) {
 		r.Post("/recreate", manager.roomRecreate)
 	})
 
-	r.Get("/events", manager.events)
 	r.Get("/docker-compose.yaml", manager.dockerCompose)
+
+	//
+	// events
+	//
+
+	r.Get("/events", manager.events)
 }
